# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FSM Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuoteListResponse" }
    post:
      operationId: createQuote
      tags: [quotes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuoteWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuoteResponse" }
  /quotes/{id}:
    patch:
      operationId: updateQuote
      tags: [quotes]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuoteWriteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuoteResponse" }
    delete:
      operationId: deleteQuote
      tags: [quotes]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
components:
...
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Quote:
      type: object
      required: [id, customerId, ownerUserId, status, subtotalCents, totalCents, createdAt]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        propertyId: { type: integer, nullable: true }
        ownerUserId: { type: integer }
        status: { type: string }
        subtotalCents: { type: integer }
        totalCents: { type: integer }
        sentAt: { type: string, format: date-time, nullable: true }
        decidedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    QuoteWriteBody:
      type: object
      required: [customerId, ownerUserId, status, subtotalCents, totalCents]
      properties:
        customerId: { type: integer }
        propertyId: { type: integer, nullable: true }
        ownerUserId: { type: integer }
        status: { type: string }
        subtotalCents: { type: integer }
        totalCents: { type: integer }
        sentAt: { type: string, format: date-time, nullable: true }
        decidedAt: { type: string, format: date-time, nullable: true }
    QuoteResponse:
      type: object
      required: [quote]
      properties:
        quote: { $ref: "#/components/schemas/Quote" }
    QuoteListResponse:
      type: object
      required: [quotes]
      properties:
        quotes:
          type: array
          items: { $ref: "#/components/schemas/Quote" }
...