# lib/api-zod/src/generated/types/quoteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Quote } from "./quote";

export interface QuoteResponse {
  quote: Quote;
}