# lib/api-zod/src/generated/types/quoteWriteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteWriteBody {
  customerId: number;
  propertyId?: number | null;
  ownerUserId: number;
  status: string;
  subtotalCents: number;
  totalCents: number;
  sentAt?: Date | null;
  decidedAt?: Date | null;
}